+++ service/src/main/java/com/epam/mediaserver/service/impl/GoodTableService.java
@@ -26,7 +26,7 @@ public class GoodTableService {
         return goods;
     }
 
-    public void delete(int idGood) throws ServiceException {
+    public void delete(Long idGood) throws ServiceException {
 
         try {
             OrderSong good = (OrderSong) SqlFactory.getOrderSongDao().getById(idGood);
