+++ web/src/main/java/com/epam/mediaserver/command/impl/delete/GoodServiceDelete.java
@@ -24,7 +24,7 @@ public class GoodServiceDelete implements Command {
         String id = request.getParameter(Parameter.PARMETER_SONG_ID);
 
         try {
-            Song song = ServiceFactory.getSongService().getById(Integer.parseInt(id));
+            Song song = ServiceFactory.getSongService().getById(Long.parseLong(id));
             ServiceFactory.getOrderUserService().remove(song);
 
         } catch (ServiceException e) {
